requirements.txt
# Microsoft Sentinel Audit Script Dependencies
azure-mgmt-loganalytics>=13.0.0
azure-mgmt-securityinsight>=1.0.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-monitor>=6.0.2
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
